refactor: remoldagem do backend com classe de tipoUsuario
