+++ Backend/app.py
@@ -6,7 +6,7 @@ routes = (os.getcwd()+'/Backend/routes')
 sys.path.append(routes)
 
 from Aluno import aluno_bp as aluno_app
-from Personal import personal_pb as personal_app
+from TipoUsuario import tipoUsuario_bp as tipoUsuario_app
 from Treinos import exercicios_bp as exercicios_app  
 
 app = Flask(__name__)
@@ -14,7 +14,7 @@ loginManager = LoginManager(app)
 app.config['SECRET_KEY'] = '@@@@@@@@'
 
 app.register_blueprint(aluno_app)
-app.register_blueprint(personal_app)
+app.register_blueprint(tipoUsuario_app)
 app.register_blueprint(exercicios_app)
 
 app.run(port = 5000, host = 'localhost', debug=True)
